Create tdsConductivityCode
Connect the signal wire from the conductivity to A1, connect VCC to power 3.3, connect Ground to Ground. When the signal pin is connected to A1 of the ADC, connect the ADC SCL line to the SCL pin of the raspberry pi, SDA to SDA pin of the Raspberry PI, and the ground and VCC to the respected pins on the Pi.